I managed to remove all errors form the new code added. However, I am still checking if it is correct.

+++ kernel/stat.h
@@ -10,8 +10,10 @@
 
 struct stat {
   int dev;     // File system's disk device
-  uint ino;    // Inode number
+  //uint ino;  
+  int ino;    // Inode number
   short type;  // Type of file
   short nlink; // Number of links to file
-  uint64 size; // Size of file in bytes
+  //uint64 size; 
+  int size; // Size of file in bytes
 };
